+++ eslint.config.mjs
@@ -154,7 +154,12 @@ export default tseslintConfig(
       'jest/expect-expect': [
         'error',
         {
-          assertFunctionNames: ['expect', 'expectIs', 'expectIsNot'],
+          assertFunctionNames: [
+            'expect',
+            'expectIs',
+            'expectIsNot',
+            'expectAllUndefined',
+          ],
           additionalTestBlockFunctions: [],
         },
       ],